Unified srvModule and server_type flags
There was no need to have both, since one can be specified with the other
(i.e., server_type = "ngamsCacheServer" is equivalent to srvModule =
"ngamsServer.ngamsCacheServer"). The new -path option accepted by the server is
now also used to add the parent directory to the path instead of manually
handling the PYTHONPATH variable on the new process.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>