ngamsDaemon stop observs process, not pidifle
There were cases in which the main thread exits fine, but there are still
hanging threads around that take longer to finish (i.e., threads serving HTTP
requests, which are currently sent to the wild and never joined properly). In
these cases the ngasmDaemon stop command finished quickly, because after the
main thread finished the pidfile was removed, and thus "stop" was happy, while
in reallity the process still existed. In turn, a rapid subsequent "start"
would fail because it wouldn't bind to the port, or any other similar problem.

This new code checks, during 'stop', that the *process* is still not around,
and also double-checks that the file doesn't exist anymore (otherwise it
removes it).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>