ngamsDaemon gaining flexibility
The ngamsDaemon now doesn't contain hardcoded command line flags and doesn't
depend on the NGAS_PREFIX variable anymore. Instead it parses its command line,
waiting for the -cfg <file> options to be present. It reads the configuration
to find out the root directory (for calculating its PID file) and the IP
address that the server exposes (when checking the service status). All
command-line parameters except [start|stop|restart|status] are passed down
verbatim to the NGAS server when it starts.

The init script is picking up on these changes, and now calculates the daemon
command line based on more options present in the init options file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>