Unifying the POST command functions, part 1
The previous state of affairs was rather poor actually, since the low-level
httpPost and httpPostUrl functions under ngamsLib were handling various data
source types (buffers in memory, files on disk, directories with files within,
etc) instead of receiving a uniform object that can be passed straight down to
httplib. By using such an object we don't need to cater for special cases and
have complex logic in the lower levels of the code to cope with high-level uses
cases. The same complexity is found in the httpGet and httpGetUrl functions,
which will be changed in the next commits.

This work is known to not be 100% complete yet. The missing piece is the
container support, for which we need to write a MIMEMultipartStream class which
iterates over a directory hierarchy in the filesystem while being read().

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>