Long waits block the server shutdown
The fact that the wait was too long meant that at shutdown the individual
delivery threads were probably waiting on this queue.get call. In turn, this
means that the main thread driving the shutdown was eventually giving up (it
uses a timeout of 10 seconds when joining threads). This means in turn that the
mechanism used to shutdown the server *process* had to eventually send a
SIGKILL to finish it off (e.g., the unit test classes that start servers, or
the ngamsDaemon class).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>