httpGet/httpGetUrl return the HTTP reponse object
This make the code much more flexible that it was before. By returning the
response we give the caller the flexibility to do whatever they want with the
content that comes back from the server (instead of having to code that
behavior in the ngamsLib module).

This also means callers should make sure they close the response object.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>