Ensuring correct shutdown of server in test
The way the server was being shutdown did not ensure the process was finished,
and therefore the new server being started by the test failed in some
situations to start up. This new code ensures that the first server is properly
shutdown before attempting to start the second one in the same test.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>