NGAS-84 installing dependencies to build the docs
This improvement has two parts: on the one hand we added the additional
installation steps on the build.sh script to install all required python
packages when indicated via a command-line flag, while on the other hand we use
this flag from the fabric script when building NGAS. These python packages are
enough to build at least the HTML documentation. The PDF version of the
document requires more system packages (latex and friends), but I'm reluctant
to add all those to the list of packages needed to build NGAS, so that's it for
now.

I also removed the html_static_path preference from the conf.py script, as in
new installations this caused problems (because we actually don't have a
_static directory in our repository).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>