NGAS-53 improving installation bits on the docs
I also renamed the "boot.sh" script to something more meaningful, and provided
more documentation about what's its purpose, both in the docs and in the
comments and messages contained in the script itself.

I also added a small intro, which can possibly be extended in the future,
and updated some instructions on how to run the client and the daemon.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>