Improving syslog handler setup
If a problem is found while setting up the syslog handler we should still try
to continue, warning the user that the syslog handler couldn't be set up. In
the future we could also allow users to pass down a specific address where to
connect to for the syslog handler.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>