Don't skip any tests now
The last run went all the way through without problems, which is probably a
pretty good sign that the recent changes to how we run the tests (always
listening on 127.0.0.1) together with the corresponding fixes really did
improve the test runs in Travis. The tests we were skipping are likely to run
correctly too, so now there's no point on holding back.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>