Reliably shut down the server during test
The way it was done before left the door open for a race condition (which we
starte seeing more reguarly in travis) where, after sending the OFFLINE and
EXIT commands, the code would start a new server on the same port *without*
making sure the previous one was fully down, thus failing. We had seen (and
fixed) the same behavior in the test_BackLogBuf_01 test of the same class, but
failed to find this one earlier.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>