Fixing bug that prevented the server to go ONLINE
This bug was creating during the unification of the time-conversion routines.
It only manifested itself when starting the server with subscripions that had
not set a value in their last_ingested column, therefore trigerring the bug. Up
until now no production system had that behavior (because their subscrptions
had been always working) and the tests didn't check that particular case.

This commit both fixes the problem, and also adds a test that reproduced the
issue.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>