"Fixing" structure conversion routine
The subscription thread uses internally a list with fields in certain positions
to maintain file information. Different db-retrieved sequences are converted
throughout the code to this internal list and used thereafter in the new
format. The conversion routine wasn't well enough written though, and was
relying on analyzing the data type of a particular field on the incoming list,
which is database-dependant. In order to quickly fix this without having to
change much of the code I've changed the identification condition to something
more reliable than that. There are some HACK notices around these places and
some TODO notices as well to let people know that this needs to be properly
fixed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>