Properly exiting the script on error
Instead of ORing the exit code of each command and executing the error function
if the exit code was not 0, I was *piping* the output of the command to the
error function. This is obviously not what I really wanted! Due to the fact
that the piped commands are treated as a whole, the exit command executed
during the piped error function execution was simply setting the exit code of
the pipelined instructions instead of exiting the script, which is why I failed
to see this problem until now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>