fabric scripts use create_venv.sh to create venv
This centralizes the creation of venvs in the create_venv.sh script instead of
having a duplicated functionality in the fabric scripts.

Note that for this to work we first need to copy the sources to the remote end
and *then* try to create the virtual environment (we were doing it in the
reverse order before because the creation of the venv didn't depend on the NGAS
sources being there).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>