Simplifying the docker_image fabric task
The code is now better organized, throwing exceptions instead of returning None
values and without much JSON parsing and printing.

A new variable is also now accepted to control the name of the final image.

Finally, I've changed the generic start_ngas_and_check_status task to stop the
server after starting it and checking it's up and running.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>