Re-organizing ngamsSql contents
First of all, most of the scripts have been moved to the "legacy" folder, which
is where they belong, and removed an sqlite3 database file and a few other
way-too-specific SQL scripts.

Secondly, I left three main scripts at the top-level directory, which take care
of creating an NGAS 9.0 database schema for mysql, postgresql and sqlite3.

Finally, under the deltas/ directory, a new structure with folders per version
first, and for database engines later, hosts the delta scripts that should be
run to migrate existing databases to newer versions of the software.

In a later commit I think I should stop using ngamsSql as a python package,
rename it simply to "sql", and move it to the top-level directory of the
repository.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>