Reconnect to DB from both parent and child procs
This ensures that the connections to the database are not shared at all between
the two processes, which produces avhoc in most cases.

Initially this patch contained only the reconnection on the newly spawned
process (the janitor proccess), who uses the database only after reconnecting.
Even with that, I was still seeing errors during my tests with MySQL on the
*parent* process right after spawning the child. I haven't tracked down exactly
what is going on there, but for the time being re-creating the connections in
the parent as well sounds like a good compromise.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>