Getting rid of global ngamsCore._uniqueNumberCount
This was a global, semaphore-guarded, increasing counter that was used to
create unique filenames during the lifetime of an NGAS server execution. Now
that we started pulling some threads into separate processes this global
counter cannot ensure we will end up with unique names based on this global
counter.

I thus:
 * Removed the counter, its semaphor and the function which incremented it
 * Removed the genUniqueFilename function that used this counter
 * Replaced all calls to the two methods above with more appropriate tempfile.*
   function calls.

Most of the new calls to tempfile.* are actually tempfile.mktemp, which is not
completely safe (it suffers from a race condition) but it's a less invasive
replacement. In the future we should move all calls to tempfile.mkstemp, which
fully ensures the file is unique, but returns an opened file descriptor for the
new file that needs to be dealt with.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>