Serializing sync command execution
This was causing problems in the ARCHIVE stress test, which ultimately caused
the server to fire up the same command from different threads at more-or-less
the same time. Process creation in python 2.7 needs to be serialized though (it
has some internal threading issues that are solved in python 3) and thus we are
using now this small lock.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>