Proper ordering of DB changes application
After changing the way we name temporary files, we accidently broke the order
in which the DB change .pickle files are processed later by the janitor thread,
who was simply sorting by filename. We cannot rely on the filename anymore to
impose an order, and thus we go for the creation date instead. In the future
this could be improved even further to use an attribute contained in the
pickled value itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>