Avoid DB lockdowns during DataCheckThread shutdown
The lockdown happened when the _finishThread method was called while a
connection to the database was still open from the DataCheckThread code.
Because _finishThread connected to the database, it was a requirement that no
database connections were held by the same thread, which happens at least once
but it could happen in more places.

The particular example that made this evident was a call to _stopDataCheckThr
in the _dumpFileInfo function. This call is made while a database cursor object
is still open, and thus the databse call inside _finishThread fails (with
sqlite3 we see an OperationalError with a "database is locked" message, but in
general we find ourselves in a deadlock if we have a pool with only one
connection).

Removing the database update from the data cheker thread is the simplest and
most generic solution, as it solves not only the particular problem we
observed, but any other that could arise in the future due to the same cause.
The database gets updated anyway after the offline procedure takes place, so
there actually was no need to do this additional update in the first place.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>