New generation of logfile handler plug-ins
This clearly separates some very specific behaviors that we may want to add in
the future from the main work of the janitor process. We also provide a first
plug-in implementation that scp's the file. This is the same logic that was
previously hardcoded in the rotated_logfiles_handler janitor module, but now as
a plug-in itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>