Removing unnecessary, duplicate objects
The tracking of disks throughout the code was very confusing, as there
were different objects keeping track of them with names, that were not
clearly named anyway. Now there is a single "disks_to_check" dictionary
that keeps track of the disks that are being checked. This dictionary is
returned by a stand-alone function too, making things easier to
understand too in the _dumpFileInfo function.

In the same spirit, the collect_files_on_disk function also isolates the
logic of walking through the mount points away from the _dumpFileInfo
function, making it easier to read.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>