Removing unnecessary code
Counting several times the number of files in the database is not a
proper way to ensure that the list of files is "static"-ish, which
seemed to be the original intention of this code. This check is not
really necessary anyway, because during check-time more inconsistencies
might come up anyway, so it's pointless to try to have a consistent
state at the beginning of the run.

There was also some exception masking code going one, so I removed that
as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>