Latest ALMA mirroring code: MIRRARCHIVE command
The MIRRARCHIVE command uses internally both the "HTTPFETCH" and
"RSYNCFETCH" modules (which are not NGAS commands on their own right,
but nevertheless they were named like if they were).

These two transport methods contact the remote NGAS server with the
RETRIEVE and RSYNC commands, respectively. The HTTPFETCH module checks
if the RETRIEVE command supports a starting offset, in which case it
uses it. This is something we haven't implemented in our RETRIEVE
command, but shouldn't be too hard. On the other hand the RSYNCFETCH
module issues an RSYNC command, which we don't have. Looking at some
ALMA configuration files it seems they are not currently using the RSYNC
transport method, but it's still good to support it and therefore I will
bring it too in a later commit.

The MIRRARCHIVE command also uses internally what seems to be a custom
copy of the ngamsDAPI module named ngamsDAPIMirroring. I've copied it
over as well, but in time we could study how to remove it and organize
this code better.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>