NGAS-91 C client should compile with brew
This is the first iteration of a solution, but not a very nice one
(because we are affecting all compilations by setting these flags
globally for the build.sh script). I also haven't tested it as I don't
have a Mac with brew with me now, but it should work.

The proper way of doing this (yet to come) involves modifying the
configure.ac script to actually look for the library/header files of
bsddb, and remove that bit from the Makefile.am file. We then need to
expose this as a user-provided command-line argument, which will allow
us to pass down that information from the fabric scripts via the
build.sh file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>