Using autoconf to look for libraries
We were previously not using autoconf's ability to look for libraries
and header files, and therefore when one of them was not present we were
failing to compile the C client (instead of getting a more obvious error
message at ./configure time).

The next step is to allow users to specify the location of the bsddb
development files so we don't need to modify the CFLAGS and LDFLAGS
environment variables.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>