NGAS-94 Default to quick_location method
The "quick" method of locating an NGAS file is way quicker than its
"normal" counterpart, as it consists simply on a database query and then
a few logical statements. The "normal" way, on the othe hand, requires
several database accesses, temporary bsddb files, and a more complex
logic. I haven't studied the reason behind this, but I suspect that most
of it can be simplified.

The "quick" method can be turned off anyway as always by specifying
quick_location=0 on the HTTP parameter set.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>