Offloading checksum calculation to external procs
This is a first, experimental version. The calculation of the checksum
should be completely offloaded into different processes, which are
created at the very beginning of the server start up (and therefore are
almost pristine).

A more proper solution would involve reworking the ngamsFileUtils'
check_file method to be more reusable and less dependent on a database
connection. This would allow us to execute that method directly in a
different process instead of passing down an executor for the particular
bit of logic that we want to isolate in a different process.

This work will also need to be merged eventually with the other commits
currently living in the datachecker-cleanup branch.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>