No need to quote out + signs
+ is only a special symbol in application/x-www-form-urlencoded data,
where it is interpreted as a space *before* further unquoting. The code
in the request properties class was doing this replacement *after*
unquoting, and on data belonging to a MIME header, which isn't affected
by this rule.

This was possibly caused by incorrectly coded client-side code that was
unnecessarily urlencoding values that were put into the
Content-Disposition header (or for that matter, on any header).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>