Optional crc32c dependency
The dependency on crcv32c (on by default) can be avoided by specifying
the NGAS_NO_CRC32C environment variable when proceeding with the
installation of the ngamsCore package. The build.sh script has a small
help mentioning this, and the fabric script also has a fabric variable
that can toggle this behaviour.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>