Triggering the datachecking load on IDLE server
Whenever the server goes to BUSY state the datacheck thread stops
immediately its load (using multiprocessing events). As soon as the
server goes IDLE again the work resumes. When the server is shutdown,
and the stop event is sent to the data checker thread, the worker
processes are also signaled for finishing its work.

This is a first iteration, which needs some testing first.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>