Unifying usage of subprocess module
There were a few os.system calls, plus a few usages of the commands
module. Now all test code uses subprocess.

Among the things I needed to change underneath was the fact that mvFile
was always changing permissions on the target file / directory to a
hardcoded 422 set (which, in the case of directories, renders the
directory unwriteable because the access/execution flag is unset, and
it's anyways simply incorrect to have this behavior coded at that
level), and adding the ability to specify a new environment when
spawning processes via execCmd.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>