Correctly count servicing threads
This code obviously suffers from a small race condition (because the
serving count is only updated a bit later, and after the thread is
actually created), but we acn live with it, and anyway will yield better
results than the previous guess (which counted all system threads,
unaware of where they had originated from).

The code sending back the response is also now actually correct, so
clients should see the HTTP code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>