Properly replying with 503 when needed
The previous version of the code was still going through the ngamsServer
and tried to serve the request, which was plain wrong. This new
implementation doesn't even create a request handler object (because
simply by creating it the request gets processed and served). Now we
write a simple HTTP reponse directly into the request socket and finish
it off.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>