Refactoring read/crc/write loop out of QARCHIVE
We moved it now into ngamsArchiveUtils. This is the first step towards
reusing such logic in other modules, like the ARCHIVE command.

The bit of logic that was moved into ngamsArchiveUtils uses a generic
file-like object as its input, containing no particular references to
the request object, etc. This way we can easily write a handful of
options on top of it to handle http-to-file transfers, file-to-file
transfers, or other data sources.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>