New GzipFile class with checksum calculation
Our new GzipFile class (which simply inherits from gzip.GzipFile)
calculates a user-provided checksum flavour in the *compressed* data as
it gets compressed. This is done by wrapping the underlying file object
on which the compressed data gets written.

The class can be used by itself, or by calling the gzip_compress method,
which creates a new instance of the class and uses it to compress data.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>