Allow DAPIs to perform checksum
When archiving incoming data, the checksum was being calculated either
on the incoming stream of data or by re-reading the final archived file
after the corresponding DAPI was executed, depending on whether the DAPI
would modify the original data or not.

This patch adds a third option: that the DAPI calculates the checksum as
it modifies the data. Depending on its internal working mechanism, the
DAPI may or may not be able to tap on a stream of modified data, which
can be used to calculate the checksum on the stream and avoid re-reading
the final file. To do this, we indicate the DAPI which checksum method
needs to be used, and the DAPI in return gives us back the checksum, if
calculated. Of course, the DAPI should not try to calculate the checksum
if it doesn't modify the data.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>