Ignore these tags
Duing tests we usually archive fits files, which end up zipped up on the
server. Ensuring the final contents of the zipped files (and therefore
their lengths and checksums) is not possible, and therefore we should
ignore in those cases (which amount to almost all cases) these tags
coming from status documents.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>