Unifying QARCHIVE and ARCHIVE
This commit finally brings together the code used by the QARCHIVE and
ARCHIVE commands. Until now they both used different code paths, which
looked very much like each other, only with small differences. There
were a few details that were better handled in one than in the other,
and I tried to keep all the "good" parts of the code in this final
version.

The common code still lives under ngamsArchiveUtils, and is invoked with
different parameters by ARCHIVE and QARCHIVE. There is still an init
method call and a data handling method call which can eventually be
merged in the future.

There is still some code replication going on by commands like LARCHIVE,
the ngamsArchiveUtils.archiveFromFile method, and the CARCHIVE command.
In the future they should all point to the same place, further reducing
the amount of code that we have to deal with.

As part of this code merging I also removed some unnecessary exception
constructions, always re-raising the original exception up. This finally
affects the error messages seen by the user, and therefore some test
reference files need to be adjusted.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>