Flexible transfer methods supported for archiving
Instead of having a hardcoded HTTP-based archiving using the incoming
request as the file descriptor, we now support archiving of files using
different transfer methods.

The first user of this new transfer mechanism is the BBCPARC command,
which now doesn't need to replicate most of the archiving boiler-plate
code. It uses the same options as those used by QARCHIVE, as the
duplicated code came originally from there.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>