checksum_info-based checksuming
The previous implementation of checksuming assumed that the initial
value of the checksum algorithm was always 0, and that at the end of it
no final adjustments were needed. This was true at least for crc32c (or
how we are using it), but is definitely not true for crc32c, when
comparing the results with those from other tools.

This commit encapsulates the initial value, the method, and the final
adjustment of each checksum variant in a named tuple, which is then
passed around and used as it corresponds.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>