BBCPARC command aware of latest bbcp developments
The latest bbcp version (that we maintain in our own git repo) has
crc32c support. NGAS can thus now use that support to transfer files
that have been checksumed with crc32c.

I've updated the tests to double-check that this is working. The tests
check the bbcp version installed in the system and adjust their
expectations accordingly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>