Properly check for bbcp presence
I can't remember what the semantics of the `which` command are, but they
clearly were not what I expected them to be (i.e., if the executable
being asked about is not there, it still seems to return 0). On the
other hand, we had skipIf(bbcp_cmd), meaning that we were skipping the
tests if (under our old understanding) bbcp WAS present, which made no
sense.

This new way of testing for bbcp should be more resilient to errors, and
should enable/disable the tests as intended.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>