NGAS-103 support for archiving event handling
This bit of code gives us the flexibility to react on archiving events.
Archiving events are generated at the end of a successful file
archiving. A small event containin the file ID and the file version is
generated and passed down to each of the event subscribers.

The server already had a bit of post-archiving logic (i.e., it triggered
the subscription thread that publishes files to subscribers). Because
this bit of logic fits nicely this event subscription mechanism, I've
moved it and now lives as a built-in subscriber.

Users can also specify their own subscribers via plug-ins specified in
the configuration file. This way even more flexibility is available.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>