Subscription test uses new archiving evt mechanism
Instead of simply waiting for files to arrive at the subscriber server,
we now place an archiving event subsriber on the subscriber server
(confusing, I know) that sends a packet to a socket we open *from* the
test. This way the test will be aware of the file being archived in the
subscriber server as it happens instead of blindly waiting for a given
amount of time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>