allow_reuse_address to always correctly bind
We might get "address already in use" errors when binding our listener
socket due to race conditions between the test case closing the listener
socket and NGAS plug-in closing the sending socket. It's safer to simply
allow to reuse the address, which should let us through every time.

This is hopefully the real reason why Travis was failing every time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>