First step towards standarizing command interface
The built-in NGAS commands had each a different entry-point method
unnecessarily. This made the caller method (in ngamsCmdHandling) way
more complex that it needed to be.

All commands now follow the same interface (which is the same that
plug-in commands already followed!). This makes the code in
ngamsCmdHandling much easier to read and follow. There is still some
special handling for a few static resources that should probably be
removed (and let them return 404 instead?). I also realized that modules
were being kept at the ngamsServer level without reason, so I removed
that as well.

The next step is to support arbitrary modules for command
implementation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>