Command plug-ins implemented in arbitrary modules
With the previous code users could not ship command plug-ins separately
from NGAS, as the internal mechanism for resolving commands was
hardcoded to look for the ngamsPlugIns.ngamsCmd_<CMD> pattern. The
ngamsPlugIns package is shipped by NGAS, meaning that even though users
could implement commands in their own modules, they had to be first
renamed to look like ngamsCmd_<CMD>, and then they had to either be
manually placed under the ngamsPlugIns package installation (not nice),
or they had to be included in the vanilla NGAS codebase (even uglier).

This new code allows users to specify in the configuration file where
commands can be found (i.e., which python module implements it, and not
necessarily the actual location on disk). A new test double-checks that
this loading mechanism indeed works.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>