crc_benchmark enhanced
Now we print the CRC both as a bit pattern, and as the integer value
stored in the NGAS database (as a string). This makes it easier to have
a clear view of what goes into the database at the end, but what the
actual checksum is (to compare with other tools).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>