Always mask crc32c result with 0xffffffff
NGAS stores the string representation of the python int value of the
checksum result (instead a more direct representation of the bytes), and
thus we need to take this precaution, which is needed to obtain
consistent results across different platforms. This is clearly
documented for the binascii.crc32 method in the python documentation,
and the same reasoning applies to the crc32c method.

The underlying reason is that different versions of Python return a
different *number*, although the underlying bit pattern is the same.
For instance, in Python 2.7 the C module interprets the crc result as a
signed integer, and expands it into a long; therefore the Python code
might see the result as a negative value. In Python 3 however this is
not the case, as the result is passed as an unsigned long when
constructing the Python integer that the user receives.

I also made a note in the crc32 variant to make it clear that this is
inconsistent across python versions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>