New 'crc32z' checksum variant supported
This is identical to the 'crc32' variant, but with the nice property
that its integer value will be the same across different versions of
python. The name 'crc32' was already taken (by the inconsistent crc32
variant), so I had to choose a different one. Since this is the
implementation found in zip files, I chose 'crc32z'.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>