Make byte-to-int conversion crc-specific
NGAS stores checksum results as the string representation of an integer
value. Because this integer value will be different depending on how the
crc variant interprets the 32 bits of the checksum value (crc32
interprets them as signed int in Python 2.7, but our crc32c ensures they
will have always an unsigned value), the task of converting bytes into
integers (like what we use in BBCPARC) should be crc-specific as well.

New tests also ensure that we correctly handle checksum values with
their most significant bit set and unset (which makes the distinction
beetween a signed or unsigned value) for all the supported variants.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>