Method-dependent checksum comparison
Different checksum methods interpret bytes differently when converting
them to integer values. By having them also be in charge of comparing
these values we can make sure these values are self-consistent.

This of course makes the code in general a bit more convoluted (because
we are calling checksum_info.equals(x, y) instead of simply x == y), but
at least it guarantees that we are doing The Right Thing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>