Making to_status reusable
This code transforms the data coming back from an HTTP response object
into an ngamsStatus object. If there is no data in the response a dummy
ngamsStatus object is created instead, with the appropriate status field
set depending on the HTTP return code.

This code was previously present in the ngamsPClient module only, but it
is needed in some other places inside the server, which is why it's
being moved to a common location.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>