Removing manual usage of httplib/urllib
We already have the ngamsHttpUtils and ngamsPClient modules, which we
should try to use in more places. Most of the plain httplib/urllib
usages were in the test modules, but there were some in the
ngamsServer and ngamsPlugIns. There are a few missing, but in modules we
know we don't use, so it's less urgent to port those.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>