get_endpoint -> get_self_endpoint
This is a better description of what this method does. This is only
meant to be used when the server needs to issue a command to itself, and
not when a host/port pair needs to be given to *other* entity that needs
to contact us. The latter case must be done either by inspecting the
request URL, like in any other web framework in python and other
languages (if within the context of serving a request), or by
aggregating the exposed address together with any available proxy
information.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>