Updating how SubscriptionDefs are processed
Users should not need to specify the full URL in the configuration file,
because that will depend on the IpAddress that is set in the Server
element (and in the NAT/reverse proxy information that we need to add to
the server as well). Thus, the SubscriberUrl attribute is marked as
deprecated, and the Command attribute (which indicates the local command
that the remote server should us when pushing data onto us) is asked for
now.

The configuration files used during unit test had the Enabled flag set
to 1 (unnecessarily), which in turn started showing lots of warning
messages because they hadn't been ported to use the new attribute yet.
This commit thus turns the Enabled flag off for all of them, and
prepares them for a new unit test that will be coming in the next commit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>