Reimplementing remote subscription creation
This new implementation is simpler and more correct. It checks correctly
if a subscription will cause a loop against ourselves, it constructs the
URL for the subscription using the information of our local endpoint
(which doesn't include yet NAT/reverse proxy support), and reacts better
to server shutdowns (uses an event instead of blindly sleeping for short
periods of time).

A new unit test also makes sure we exercise this new code properly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>