New table and column definitions
Two new elements have been defined: a new ngas_subscr_delivery_queue
table (which will hold all the deliveries that are pending on the
system), and a new column on the ngas_subscribers table to indicate
whether a given subscriber is active or not.

Unit tests make sure the new table gets emptied every time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>