Removing old subscription queries, adding new ones
Given how we designed the new subscription delivery system, we don't
need anymore much of the old queries which interacted with the
ngas_subscr_queue and ngas_subscr_back_log tables. Instead, we now focus
solely on the ngas_subscr_delivery_queue table, which should maintain a
fairly low number of entries at any given time. The exceptions is when
there are failures on the remote end, in which case entries will
accummulate, very similarly to how they did on the old back_log table.

Also, the old queries touching the ngas_subscribers table have been
improved a bit for easier reading and maintenance.

Note that some of the new queries use the new transaction() class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>