New transaction and cursor2 objects
With the new transaction object we can now perform multiple SQL queries
within the same transaction, which is something we should have
implemented some time ago already.

With the new cursor2 object we can now call cursor.fetch(x) and simply
iterate over the results (which will be retrieved in batches of x
elements), because fetch() returns a generator. We are now optionally
using this new cursor2 object as the return type for the getFileSummary2
query, which will be used in the new subscription thread implementation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>