First implementation of new subscription thread
This implementation is much simpler than the old one. It removes any
references to the job:// scheme that was at some point supported by
URLs, it simplifies the interaction between the main subscription driver
thread and the worker threads (they are more synchronous now), and
uses the new delivery queue table to simplify the tracking of what needs
to be done.

The simplicity of this code basically takes down the number of lines of
the file from ~1400 to ~200, while doing the same job in a more
efficient way.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>