Re-implementing UNSUBSCRIBE command
Its logic now is as simple as removing the corresponding subscription
row from the subscribers database table, and removing its associated
deliveries from the delivery table. Any delivery that is already
scheduled to happen during the current major subscription delivery cycle
will effectively be delivered, but we can live with that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>